+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Added `BundleInserter` trait
 - Changed: generation runner plugins( simple & debug) now emit bevy `Event` to signal generations and resets: `GridGeneratedEvent`, `GenerationResetEvent` and `NodesGeneratedEvent`
 - Added `ProcGenSpawnerPlugin` to spawn the node entities with observers
+- Added `ProcGenSimplePlugins` and `ProcGenDebugPlugins` plugin groups (runner+spawner)
 - Renamed the `default-assets-bundle-spawners` cargo feature to `default-bundle-inserters`
 
 - Changed: remove `bevy_mod_picking_dependency` and use bevy 0.15 internal picking features